upload procedure doesn't preserve bootloader
Withdrawing the .hex and  upload script.  The process as is somehow clobbers the bootloader.  Meaning, unless you want to burn the bootloader again, this is a 1-time upgrade, not what we want here.  We will update the process to preserve the bootloader and re-publish the hex file and script.  Thanks all.